requirements.txt
coloredlogs
python-telegram-bot>=13.1
requests>=2.24.0
python-dotenv